+++ shared/util/tf/tf_parser.go
@@ -1,4 +1,4 @@
-package util
+package tf
 
 import (
 	"fmt"
