+++ shared/util/tf/tf_parser_test.go
@@ -1,4 +1,4 @@
-package util
+package tf
 
 import (
 	"testing"
